requirements.txt
langsmith==0.1.142
pandas
matplotlib
langgraph==0.2.45
langchain-core==0.3.15
fastapi
uvicorn
langchain-community==0.3.5   
langgraph-checkpoint-sqlite
python-dotenv==1.0.1
termcolor
unidecode
psycopg==3.2.3
psycopg-pool==3.2.3
langgraph-checkpoint-postgres==2.0.2
langchain-together
ipython